Rename and refactor pages
