+++ frontend/src/pages/teacherDashboard/teacherClasses/editClass/additionalClassSettings/RapidRouterTabTitles.tsx
