+++ frontend/src/pages/teacherDashboard/teacherClasses/editClass/additionalClassSettings/RapidRouterTabs.tsx
